[dev] 网络请求库开发

+++ app-core/src/main/java/com/pandaq/appcore/framework/base/IBaseContract.java
@@ -33,7 +33,7 @@ public interface IBaseContract {
          * @param errCode 错误码
          * @param errMsg  错误信息
          */
-        void onError(int errCode, String errMsg);
+        void onError(long errCode, String errMsg);
 
         /**
          * 结束回调
